feat: replace ScheduleTimeForm with ScheduleDateTimeForm for unified date and time input handling
